Integrate Spring REST Docs into MockMvc-based tests
This commit integrates Spring REST Docs into the tests that use MockMvc.
To allow REST Docs’ JUnit Rule to be used, the logic that creates and
configures the MockMvc instance has been moved into the @Before method
of InjectedMockContextTest. This allows the JUnitRestDocumentation rule
to be declared on a test class (where it needs to be for JUnit to find
it) and for it to then be used when configuring the MockMvc instance to
use REST Docs. The web application context is still only created once,
thereby retaining the performance improvements that were implemented in
48027bac.